+++ package.json
@@ -23,5 +23,8 @@
   "homepage": "https://github.com/TankovaE/my-incredible-project#readme",
   "devDependencies": {
     "nodemon": "^2.0.6"
+  },
+  "dependencies": {
+    "express": "^4.17.1"
   }
 }
