+++ views/about.html
@@ -3,10 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>About</title>
+    <title>Document</title>
 </head>
 <body>
     <h1>About page</h1>
-    <a href="/">Home page</a>
 </body>
 </html>
