+++ views/index.html
@@ -3,15 +3,10 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home page</title>
+    <title>Document</title>
 </head>
 <body>
-    <h1>Form</h1>
-    <form method="post" action="/">
-        <input name="title" type="text"/>
-        <button type="submit">send</button>
-    </form>
-    <a href="/about">About page</a>
-    <a href="/api/users">API (get users)</a>
+    <h1>Home page</h1>
+    <a href="/about">about page</a>
 </body>
 </html>
